requirements.txt
uvicorn
fastapi[standard]
mysql-connector-python